(change) product stock type changed to Integer

+++ src/main/java/com/utn/ProgIII/model/Product/Product.java
@@ -37,5 +37,5 @@ public class Product {
 
     private Float profitMargin;
 
-    private Float stock;
+    private Integer stock;
 }
